backend/requirements.txt
fastapi>=0.115.0
uvicorn==0.27.0
chess==1.11.2
python-multipart==0.0.6
Pillow==10.2.0
google-genai==1.9.0
httpx>=0.28.1
pytest==8.1.1
pytest-asyncio==0.23.6
